Extract code from external-dns and apply LICENSE header.
