+++ main.go
@@ -2,15 +2,10 @@ package main
 
 import (
 	"fmt"
-	"net"
-	"net/http"
-	"os"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/gophercloud/gophercloud"
-	"github.com/gophercloud/gophercloud/openstack"
 	"github.com/gophercloud/gophercloud/openstack/dns/v2/recordsets"
 	"github.com/gophercloud/gophercloud/openstack/dns/v2/zones"
 
@@ -19,10 +14,10 @@ import (
 
 	"github.com/jetstack/cert-manager/pkg/acme/webhook/apis/acme/v1alpha1"
 	"github.com/jetstack/cert-manager/pkg/acme/webhook/cmd"
-	"github.com/kubernetes-incubator/external-dns/pkg/tlsutils"
 )
 
 const GroupName = "acme.syseleven.de"
+
 
 func main() {
 	cmd.RunWebhookServer(GroupName,
@@ -35,7 +30,6 @@ type designateDNSProviderSolver struct {
 }
 
 func (c *designateDNSProviderSolver) Name() string {
-	log.Debugf("Name() called")
 	return "designateDNS"
 }
 
@@ -133,89 +127,6 @@ func (c *designateDNSProviderSolver) Initialize(kubeClientConfig *rest.Config, s
 
 	c.client = cl
 	return nil
-}
-
-// copies environment variables to new names without overwriting existing values
-func remapEnv(mapping map[string]string) {
-	for k, v := range mapping {
-		currentVal := os.Getenv(k)
-		newVal := os.Getenv(v)
-		if currentVal == "" && newVal != "" {
-			os.Setenv(k, newVal)
-		}
-	}
-}
-
-// returns OpenStack Keystone authentication settings by obtaining values from standard environment variables.
-// also fixes incompatibilities between gophercloud implementation and *-stackrc files that can be downloaded
-// from OpenStack dashboard in latest versions
-func getAuthSettings() (gophercloud.AuthOptions, error) {
-	remapEnv(map[string]string{
-		"OS_TENANT_NAME": "OS_PROJECT_NAME",
-		"OS_TENANT_ID":   "OS_PROJECT_ID",
-		"OS_DOMAIN_NAME": "OS_USER_DOMAIN_NAME",
-		"OS_DOMAIN_ID":   "OS_USER_DOMAIN_ID",
-	})
-
-	opts, err := openstack.AuthOptionsFromEnv()
-	if err != nil {
-		return gophercloud.AuthOptions{}, err
-	}
-	opts.AllowReauth = true
-	if !strings.HasSuffix(opts.IdentityEndpoint, "/") {
-		opts.IdentityEndpoint += "/"
-	}
-	if !strings.HasSuffix(opts.IdentityEndpoint, "/v2.0/") && !strings.HasSuffix(opts.IdentityEndpoint, "/v3/") {
-		opts.IdentityEndpoint += "v2.0/"
-	}
-	return opts, nil
-}
-
-// authenticate in OpenStack and obtain Designate service endpoint
-func createDesignateServiceClient() (*gophercloud.ServiceClient, error) {
-	opts, err := getAuthSettings()
-	if err != nil {
-		return nil, err
-	}
-	log.Infof("Using OpenStack Keystone at %s", opts.IdentityEndpoint)
-	authProvider, err := openstack.NewClient(opts.IdentityEndpoint)
-	if err != nil {
-		return nil, err
-	}
-
-	tlsConfig, err := tlsutils.CreateTLSConfig("OPENSTACK")
-	if err != nil {
-		return nil, err
-	}
-
-	transport := &http.Transport{
-		Proxy: http.ProxyFromEnvironment,
-		DialContext: (&net.Dialer{
-			Timeout:   30 * time.Second,
-			KeepAlive: 30 * time.Second,
-		}).DialContext,
-		MaxIdleConns:          100,
-		IdleConnTimeout:       90 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-		TLSClientConfig:       tlsConfig,
-	}
-	authProvider.HTTPClient.Transport = transport
-
-	if err = openstack.Authenticate(authProvider, opts); err != nil {
-		return nil, err
-	}
-
-	eo := gophercloud.EndpointOpts{
-		Region: os.Getenv("OS_REGION_NAME"),
-	}
-
-	client, err := openstack.NewDNSV2(authProvider, eo)
-	if err != nil {
-		return nil, err
-	}
-	log.Infof("Found OpenStack Designate service at %s", client.Endpoint)
-	return client, nil
 }
 
 func quoteRecord(r string) string {
